Bugfix for runtime error for bbb cross-compile build.
Refactor of project_1 report code caused a runtime failure when cross compiling
from ubuntu to the bbb. Moving the definition of set_size as a constant into
project_1_report results in a runtime failure when cross compiling for the BBB
with a debug build. Cross compiling a release build is fine. Both release and
native builds on the BBB are fine. The runtime failure is that the constant
changes value from 32 to 17 between the last C instruction of initialize_set_1
and the creation of set_2 in main....? Some sort of stack corruption? Can't see
anything useful in the cross compiled executable when
running it in gdb.

Testing:
  ubuntu:
    make test
    make release=1 test
    make platform=bbb all, transfer to bbb
    make platform=bbb release=1 all, transfer to bbb