Initial implementation of circular buffer.
Initial implementation of a generic circular buffer. Buffer is an opaque type
with access methods. It is initialized with a number of items to store and the
size of each item in bytes. Different buffers can be created to store different
length items, .i.e 1, 2, or 4 bytes, but each buffer should store items of a
fixed size.

Testing:
  macOS: 3 circular buffer tests are skipped. All others pass.