Implementation of data analysis routines for logger.
Cleanup logger, use a typedef for the item size, separate creation of a
log item from setting data. Assign the max number of bytes to the log
item buffer.

Add routines to process the data and store results in a summary object.

Add simple python script to process logs and print them in a readable
format.

Rename the output executable.

Testing:
  macOS - debug and release builds - all tests pass, executable runs.
  macOS x frdm - debug and release builds compile.