Bugfix in logger transmit logic with polling.
Turning on old debug output for the report revealed a logic bug in the
transmit function. When using polling all data had to be copied to the
transmit circular buffer before any was removed. If the data size
exceeded the circular buffer size, the code would hang in an infinite
loop. This wasn’t a problem when using interrupts.

Consolidated some debugging code.

Testing:
  macOS - debug and release builds ok, tests pass
  macOS x frdm - debug and release builds compile. debug build transfer
                 to frdm runs as expected.