Bugfix determining profiling timer count with overflow
Logic error setting up the profiler to account for overflow. Need to use the
max timer value, not the number of bits, when calculating time with overflow.
Refactor interface to pass start and end time to elpased time so we can unit
test the elapsed time computations. Add unit tests for overflow.

Testing:
  macOS - debug build passes tests
  macOS x frdm - debug build compiles