Change the toggling of leds with pwm.
Can't use gpio control to toggle pins on and off when the pins are muxed to the
TPM. Set the trigger value to zero so it will never trigger inorder to disable
the pin. Any value larger than 0 is enabled. When toggled on, start at 1/8 of
max so increase and decrease will have a visible effect.

Testing:
  macOS - debug build passes tests
  macOS x frdm - debug build compiles.
  fedora x frdm - debug build compiles, commands over uart work as expected.
  fedora - debug build passes test.