Port to beagle bone black as embedded system.
Port code base to the beagle bone black as an embedded system instead of treating it as a host system. dma, timers, rtc, uart are just copies of the host code. spi uses spidev to access hardware.

spi appears to work, but the nrf24 code is not enabled on the beagle
bone.

Testing:

    bbb - compile and run on board, post tests run, but have some
          differences from the baseline.

    host - not tested

    frdm-kl25z - not tested